# resourceflow-cli/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ResourceFlowApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/admin/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List notification history with optional filters, newest first.\nGET /api/admin/notifications?venueId=1\u0026amp;type=BookingCreated\u0026amp;unreadOnly=true\u0026amp;page=1\u0026amp;pageSize=20\nvenueId is optional \u2014 omit to see all venues.\ntype values: BookingCreated | BookingCancelled | VenueNearlyFull",
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "pattern": "^-?(?:0|[1-9]\\d*)$",
                  "type": [
                    "integer",
                    "string"
                  ],
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "pattern": "^-?(?:0|[1-9]\\d*)$",
                  "type": [
                    "integer",
                    "string"
                  ],
                  "format": "int32"
                }
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "type": "array",
                "items": {
                  "pattern": "^-?(?:0|[1-9]\\d*)$",
                  "type": [
                    "integer",
                    "string"
                  ],
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/notifications/unread-count": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Unread count badge. Omit venueId to get total across all venues.\nGET /api/admin/notifications/unread-count\nGET /api/admin/notifications/unread-count?venueId=1",
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/notifications/{id}/read": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark a single notification as read.\nPATCH /api/admin/notifications/{id}/read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/notifications/read-all": {
      "patch": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark all notifications for a venue as read.\nPATCH /api/admin/notifications/read-all?venueId=1",
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/push/subscribe": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Register a browser push subscription. One call covers every location \u2014 an admin\naccount is not scoped to one, so neither is the subscription.\nPOST /api/admin/push/subscribe\nBody: { endpoint, p256dh, auth, userAgent? }",
        "description": "A \u0060venueId\u0060 query parameter is accepted and ignored so a browser still\nrunning the previous frontend build keeps working against a deployed backend.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscribeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscribeRequest"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Remove a push subscription (logout / permission revoked).\nDELETE /api/admin/push/subscribe\nBody: \u0022https://fcm.googleapis.com/...\u0022",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/notifications/all": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "venueId",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/notifications/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PushSubscribeRequest": {
        "required": [
          "endpoint",
          "p256dh",
          "auth"
        ],
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "p256dh": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          },
          "userAgent": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}